It works now, calling and retrieving results from hierarchy
But still cannot directly put the truth table there to test it because I haven't implemented a way to assign inputs directly from the top object...

Wondering if I should write another Class just for handling it over the Driver Code side, over the main function, instead of fiddling with the existing code in the LogicGate Class.

Really satisfied with the result, and really thank the guy from discord telling me why I couldn't do it last time, because I didn't reference the object so the objects were copied, hence no results could be modified.

This can be built into a really handy framework to handle similar stuff that requires calling deep into the hierarchy using objects, cuz it's more flexible and you can retrieve a large varieties of information from an object, so it's not just like a recursive function that works like a tree.

I will try to implement memoization and other dynamic programming tricks into similar frameworks to see if it helps with the performance a lot.

For now, I need to make the truth table works on this before moving onto other stuff.