The UseGate Class finally works, other than...
other than UseGate::clear(), if you clear the result using this function, and use the same object to calculate again, the output will be stacked on top of the previous output, the output is correct i think, but the previous output is kept there.

I need to fix this before going onto testing it with a large truth table.